+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Bug Fixes/Enhancements
+
+- **helm**: Fix `vendor --prune` deleting charts with a custom directory
+  **[#717](https://github.com/grafana/tanka/pull/717)**
+
 ## 0.22.0 (2022-06-03)
 
 ### Features